Renamed TileLayer::width/height to columns/rows

+++ src/Scenes/Scene.cpp
@@ -34,7 +34,7 @@ namespace Tristeon
 		}
 	}
 
-	Layer* Scene::layerAt(const unsigned int& index) const
+	Layer* Scene::layerAt(const unsigned long long& index) const
 	{
 		if (index < 0 || index > _layers.size() - 1)
 			throw std::invalid_argument("Index must be more than 0 and less than the amount of layers");
